requirement.txt
"""
Libraries required for this project
"""

asgiref==3.6.0
astroid==2.15.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
Pillow==9.4.0
platformdirs==3.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pylint==2.16.3
pyngrok==5.2.1
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0
wrapt==1.15.0
